Add acoustic blinky
This commit adds an acoustic equivalent to ofx_Blinky.
The acoustic Blinky works on top of the visual version.
Different modes can be set where sound is off (default),
a beep occurs every time the Blinky is switched on and
beeps occur when the Blinky is switched either on or off.
The beep can be switched to be high or low (default).